Add files via upload
This is part of the Final Project for my Data Bootcamp course. It compares performances between various ways of predicting stock movements. Included are the FFF5 model, SARIMA, and LSTM. Over the break, I plan to add the Black-Scholes model to make this project more useful in generating a profit.